requirements.txt
faker
pymongo
kafka-python
pandas